requirements.txt
streamlit
openai
pypdf
python-dotenv
numpy